Refactor : constant 위치 변경
<description>
단순히 상수들의 모음이기 때문에 controller에 있을 필요가 없다고
생각했다.
